Upgrade Spring Boot version and streamline test methods
Updated the Spring Boot version from 3.2.0 to 3.2.3 in build.gradle.kts for enhanced performance and compatibility. Additionally, the test methods in ProductRestControllerNaiveTest.java were modified to streamline the code, specifically by combining the expectStatus() validation with its expected outcome on the same line for clarity and brevity. Also, the distributionUrl was changed in gradle-wrapper.properties, upgrading from Gradle 8.4 to 8.5.